larapus

+++ resources/views/datatable/_action.blade.php
@@ -1,4 +1,6 @@
-{!! Form::model($model, ['url' => $form_url, 'method' => 'delete', 'class' => 'form-inline'] ) !!}
-<a href="{{ $edit_url }}">Ubah</a> |
-{!! Form::submit('Hapus', ['class'=>'btn btn-xs btn-danger']) !!}
+{!! Form::model($model, ['url' => $form_url, 'method' => 'delete', 'class' => 'form-inline js-confirm',
+                 'data-confirm' => $confirm_message] ) !!}
+    <a class="btn btn-warning" href="{{ $edit_url }}">Ubah</a> &nbsp; || &nbsp;
+    <a class="btn btn-info" href="{{ $show_url }}">Lihat</a> &nbsp; || &nbsp; 
+    {!! Form::submit('Hapus', ['class'=>'btn btn-xs btn-danger']) !!}
 {!! Form::close()!!}
